feat: Make it possible to disable by_alias for schema generation (#33)
This build upon the logic added by @mortenkrane in
https://github.com/kolonialno/django-api-decorator/pull/10

### Current workings
The main point for adding this is to make the feature _complete_. Per
now, if you use a `CamelModel`, as shown in the tests, combined with
setting `serialize_by_alias=True` at your view, you will be able to
automatically serialize the outgoing response and validate the incoming
body with the pydantic model 👍🏻
```py
class MyCamelCasePydanticModel(BaseModel):
    model_config = ConfigDict(
        alias_generator=camel_case,
        populate_by_name=True,
    )
```

### Issue
If you start using a`CamelCasePydanticModel` in your view, the
`openapi.py` generator that creates the schema will always use the
alias, no matter if the view has the `serialize_by_alias=True` set or
not. This is because `json_schema()` has `by_alias=True` as[ default per
the
docs](https://docs.pydantic.dev/latest/concepts/json_schema/#miscellaneous-notes-on-json-schema-generation).
So this means if you use `CamelCasePydanticModel` on a view, the
**runtime** and **schema** can be out of sync.

> The schema is generated by default using aliases as keys, but it can
be generated using model property names instead by calling
[model_json_schema()](https://docs.pydantic.dev/latest/api/base_model/#pydantic.BaseModel.model_json_schema)
or
[model_dump_json()](https://docs.pydantic.dev/latest/api/base_model/#pydantic.BaseModel.model_dump_json)
with the by_alias=False keyword argument.

### Propose solution
Add `API_DECORATOR_GENERATE_SCHEMA_BY_ALIAS` to allow the user to turn
it off when generating schema.