feat: use serialization mode when dumping response schemas (#41)
Computed fields are part of the response being returned anyways, so it
should be part of the schema.

I decided to only target response schemas, and leave request/query
params as validation. It might be up to each types generator if you have
to add readOnly fields to the request payload if they're present, but
they are not validated on request, and therefore not needed in the
schema.